chore(services): Move server boilerplate to boltzmann
